pyproject.toml
[tool.poetry]
name = "cirec-01-llm-falcon"
version = "0.1.0"
description = "RAG system"
authors = ["Ci-re <cirecleen@gmail.com>"]
readme = "README.md"
packages = [{include = "scripts"}]
include = ["scripts/*", "datasets/*"]

[tool.poetry.dependencies]
python = "^3.12"
datasets = "3.2.0"
elasticsearch = "8.17.0"
google-generativeai = "0.8.3"
grpcio-status = "1.69.0"
ipykernel = "6.29.5"
minsearch = "0.0.2"
nltk = "3.9.1"
openai = "1.59.8"
python-dotenv = "1.0.1"
torch = {version = "^2.5.1+cpu", source = "pytorch_cpu"}
torchvision = {version = "^0.20.1+cpu", source = "pytorch_cpu"}
sentence-transformers = "^3.3.1"


[[tool.poetry.source]]
name = "pytorch_cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
